api/requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pillow==10.0.0
httpx==0.24.1
scipy==1.10.1
scikit-image==0.20.0